3-/4-bit color; concise, better, diff.
When &t_Co is 8 (in tty, for example), this colorscheme stays true to
its fox-like demeanor. Also, if the user sets their 4-bit palette to the
palette in vulpo.vim, then when &t_Co is set to 16, it will still look
good. Of course, altering the 4-bit color palette will have varying
results and will probably look bad.

Each color is set in one line using the flexible Hi command. The first
parameter is the group name. The second through fourth parameters are
used for gui-vim, when &termguicolors is set, or when &t_Co is either >=
256 or >= 16. The fifth through seventh parameters are for when &t_Co is
less than 16 (like in the tty).

Statusline changed to be light gray and white for active and slightly
darker gray and yellow for inactive bars. This is more readable.

Better support for diff also added (DiffAdded, DiffRemoved). Regular
diffs still look great.

A few links were added for C and vim filetypes, which is not ideal, but
it makes them look drastically better (subjectively). These can be
disabled by letting g:vulpo_ft_mods = 0. TODO: add this to the readme.

The code needs to be documented a little bit more soon so that it can
still be maintainable. Otherwise, it's basically complete.