past-proof map(), Title underline, NonText dark
map() in older versions of 7.4 only accepts a string. Thus, to increase
the percentage of machines this colorscheme will work on, I just changed
it from `{ key,val -> key.':'.val}` syntax to `"v:key.':'.v:val"`
syntax.

Title now has an underline, and NonText is now dark since it's only for
@ and ~ (cut-off line or EndOfBuffer).

I also started liking ts=2 sts=2. Thus, I reindented.

Modeline replaced with cole=0 so that there is no concealment in this
buffer and syn=off so that it loads *much* faster.